fix tests for possible concurrency

+++ tests/_tools.js
@@ -15,7 +15,7 @@ export const fileCleanup = (testOutputDirName, fileNamePartial) => {
             if (err) throw err
 
             for (const file of files) {
-                if (file.includes(fileNamePartial)) {
+                if (file.includes(fileNamePartial) && fs.existsSync(file)) {
                     fs.unlink(file, err => {
                         if (err) throw err
                     })
